Use kubernetes client package
